+++ packages/twenty-front/src/modules/dashboards/widgets/graph/components/GraphWidgetBarChart.tsx
