+++ packages/twenty-front/src/modules/dashboards/widgets/graph/components/GraphWidgetGaugeChart.tsx
