+++ packages/twenty-front/src/modules/dashboards/widgets/graph/components/GraphWidgetNumberChart.tsx
