+++ packages/twenty-front/src/modules/dashboards/widgets/graph/components/__stories__/GraphWidgetGaugeChart.stories.tsx
