+++ packages/twenty-front/src/modules/dashboards/widgets/graph/types/GraphColorRegistry.ts
