+++ packages/twenty-front/src/modules/dashboards/widgets/graph/types/GraphColorScheme.ts
