+++ packages/twenty-front/src/modules/dashboards/widgets/graph/utils/createGradientDef.ts
