+++ packages/twenty-front/src/modules/dashboards/widgets/graph/utils/createGraphColorRegistry.ts
