+++ packages/twenty-front/src/modules/dashboards/widgets/graph/utils/graphFormatters.ts
